+++ src/services/TicketsService.php
@@ -9,6 +9,7 @@
 namespace ether\bookings\services;
 
 use craft\base\Component;
+use ether\bookings\elements\BookedTicket;
 use ether\bookings\models\Ticket;
 use ether\bookings\records\TicketRecord;
 
@@ -40,6 +41,27 @@ class TicketsService extends Component
 			return null;
 
 		return Ticket::fromRecord($record);
+	}
+
+	public function getTicketTableElement ($eventId, $slot = null)
+	{
+		$query = BookedTicket::find();
+		\Craft::configure(
+			$query, [
+			'bookingId' => $eventId,
+			'startDate' => $slot,
+			'limit'     => null,
+		]);
+
+		return BookedTicket::indexHtml(
+			$query,
+			[],
+			['mode' => 'table'],
+			'',
+			null,
+			true,
+			false
+		);
 	}
 
 }